Enhance Sign Handling

+++ src/Crypto/Hash/CRC16_CPP.cpp
@@ -2,22 +2,22 @@
 
 extern "C" void CRC16_InitTable(UInt16 *tab, UInt16 polynomial)
 {
-	Int32 i = 256;
-	Int32 j;
+	UInt32 i = 256;
+	UInt32 j;
 	UInt16 v;
 	while (i-- > 0)
 	{
-		v = i << 8;
+		v = (UInt16)(i << 8);
 		j = 8;
 		while (j-- > 0)
 		{
 			if (v & 0x8000)
 			{
-				v = (v << 1) ^ polynomial;
+				v = (UInt16)((v << 1) ^ polynomial);
 			}
 			else
 			{
-				v = (v << 1);
+				v = (UInt16)(v << 1);
 			}
 		}
 		tab[i] = v;
@@ -26,21 +26,21 @@ extern "C" void CRC16_InitTable(UInt16 *tab, UInt16 polynomial)
 	i = 0;
 	while (i < 256)
 	{
-		tab[256  + i] = (tab[0    + i] << 8) ^ tab[tab[0    + i] >> 8];
-		tab[512  + i] = (tab[256  + i] << 8) ^ tab[tab[256  + i] >> 8];
-		tab[768  + i] = (tab[512  + i] << 8) ^ tab[tab[512  + i] >> 8];
-		tab[1024 + i] = (tab[768  + i] << 8) ^ tab[tab[768  + i] >> 8];
-		tab[1280 + i] = (tab[1024 + i] << 8) ^ tab[tab[1024 + i] >> 8];
-		tab[1536 + i] = (tab[1280 + i] << 8) ^ tab[tab[1280 + i] >> 8];
-		tab[1792 + i] = (tab[1536 + i] << 8) ^ tab[tab[1536 + i] >> 8];
-		tab[2048 + i] = (tab[1792 + i] << 8) ^ tab[tab[1792 + i] >> 8];
-		tab[2304 + i] = (tab[2048 + i] << 8) ^ tab[tab[2048 + i] >> 8];
-		tab[2560 + i] = (tab[2304 + i] << 8) ^ tab[tab[2304 + i] >> 8];
-		tab[2816 + i] = (tab[2560 + i] << 8) ^ tab[tab[2560 + i] >> 8];
-		tab[3072 + i] = (tab[2816 + i] << 8) ^ tab[tab[2816 + i] >> 8];
-		tab[3328 + i] = (tab[3072 + i] << 8) ^ tab[tab[3072 + i] >> 8];
-		tab[3584 + i] = (tab[3328 + i] << 8) ^ tab[tab[3328 + i] >> 8];
-		tab[3840 + i] = (tab[3584 + i] << 8) ^ tab[tab[3584 + i] >> 8];
+		tab[256  + i] = (UInt16)((tab[0    + i] << 8) ^ tab[tab[0    + i] >> 8]);
+		tab[512  + i] = (UInt16)((tab[256  + i] << 8) ^ tab[tab[256  + i] >> 8]);
+		tab[768  + i] = (UInt16)((tab[512  + i] << 8) ^ tab[tab[512  + i] >> 8]);
+		tab[1024 + i] = (UInt16)((tab[768  + i] << 8) ^ tab[tab[768  + i] >> 8]);
+		tab[1280 + i] = (UInt16)((tab[1024 + i] << 8) ^ tab[tab[1024 + i] >> 8]);
+		tab[1536 + i] = (UInt16)((tab[1280 + i] << 8) ^ tab[tab[1280 + i] >> 8]);
+		tab[1792 + i] = (UInt16)((tab[1536 + i] << 8) ^ tab[tab[1536 + i] >> 8]);
+		tab[2048 + i] = (UInt16)((tab[1792 + i] << 8) ^ tab[tab[1792 + i] >> 8]);
+		tab[2304 + i] = (UInt16)((tab[2048 + i] << 8) ^ tab[tab[2048 + i] >> 8]);
+		tab[2560 + i] = (UInt16)((tab[2304 + i] << 8) ^ tab[tab[2304 + i] >> 8]);
+		tab[2816 + i] = (UInt16)((tab[2560 + i] << 8) ^ tab[tab[2560 + i] >> 8]);
+		tab[3072 + i] = (UInt16)((tab[2816 + i] << 8) ^ tab[tab[2816 + i] >> 8]);
+		tab[3328 + i] = (UInt16)((tab[3072 + i] << 8) ^ tab[tab[3072 + i] >> 8]);
+		tab[3584 + i] = (UInt16)((tab[3328 + i] << 8) ^ tab[tab[3328 + i] >> 8]);
+		tab[3840 + i] = (UInt16)((tab[3584 + i] << 8) ^ tab[tab[3584 + i] >> 8]);
 		i++;
 	}
 }
@@ -72,7 +72,7 @@ extern "C" UInt16 CRC16_Calc(const UInt8 *buff, UOSInt buffSize, UInt16 *tab, UI
 	}
 	while (buffSize--)
 	{
-		v = tab[(v >> 8) ^ (*buff)] ^ (v << 8);
+		v = (UInt16)(tab[(v >> 8) ^ (*buff)] ^ (v << 8));
 		buff++;
 	}
